♻️ refactor: move reindex runtime out of database
