+++ .agents/skills/full-text-search/SKILL.md
@@ -28,8 +28,9 @@ router/service -> createFtsSearchRepo -> FtsSearchRepo -> selected backend -> ex
 - `packages/database/src/schemas/ftsSearchSyncOutbox.ts` and
   `packages/database/src/repositories/ftsSearchSyncOutbox/` own durable change capture, claims,
   retries, dead letters, leases, revision fences, and capture-definition validation.
-- `packages/database/src/repositories/ftsSearchReindex/` and `scripts/elasticsearchReindex/` own the
-  resumable full backfill. `apps/server/src/services/ftsSearchSync/` and
+- `scripts/elasticsearchReindex/` owns the resumable full-backfill command and its operational
+  runtime. Shared database source queries and document construction remain in
+  `packages/database/src/repositories/ftsSearchDocument/`. `apps/server/src/services/ftsSearchSync/` and
   `scripts/elasticsearchSync/` own continuous incremental draining.
 - `packages/env/src/ftsSearch.ts` owns generic Elasticsearch environment variables.
