+++ scripts/elasticsearchReindex/runtime/__tests__/auditLogger.test.ts
@@ -5,7 +5,7 @@ import path from 'node:path';
 
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
-import { FtsSearchReindexFileLogger, summarizeFtsSearchReindexError } from '../logger';
+import { FtsSearchReindexFileLogger, summarizeFtsSearchReindexError } from '../auditLogger';
 
 let stateDirectory: string;
