+++ scripts/elasticsearchReindex/runtime/index.ts
@@ -1,0 +1,4 @@
+export * from './auditLogger';
+export * from './checkpointRepository';
+export * from './elasticsearchClient';
+export * from './reindexService';
