+++ tsconfig.json
@@ -43,7 +43,6 @@
     "apps/mobile/**",
     "apps/share/**",
     "apps/workbench/**",
-    "scripts/**",
     "tmp",
     "temp",
     ".temp",
@@ -62,6 +61,9 @@
     "tests/**/*.tsx",
     "apps/server/**/*.ts",
     "apps/server/**/*.tsx",
+    "scripts/elasticsearchReindex/**/*.ts",
+    "scripts/elasticsearchSync/**/*.ts",
+    "scripts/installFtsSearchSyncCapture/**/*.ts",
     "*.ts",
     "*.mts",
     "*.cts",
